+++ self-managed/ops/scenarios/04_service_mesh_access/03_start_consul_server.sh
@@ -33,18 +33,32 @@ header1 "Starting Consul server"
 ##########################################################
 header2 "Generate Consul servers configuration"
 
+## This needs to be exported here to work inside the script.
+## [bug] [conf] Configuration generated does not work for different dc.domain
+export CONSUL_DOMAIN=${DOMAIN}
+export CONSUL_DATACENTER=${DATACENTER}
+export CONSUL_SERVER_NUMBER=${SERVER_NUMBER}
+
 ## [cmd] [script] generate_consul_server_config.sh
 execute_supporting_script "generate_consul_server_config.sh"
 
 ##########################################################
 header2 "Copy Consul servers configuration files"
 
-## ~todo make all servers discoverable from bastion host
+## [x] todo make all servers discoverable from bastion host
 for i in `seq 0 "$((SERVER_NUMBER-1))"`; do
-  
-  ## [mark] this thing is ugly. Debug and check paths
-  log "Remove pre-existing configuration and stopping pre-existing Consul instances"
-  remote_exec consul-server-$i "sudo rm -rf ${CONSUL_CONFIG_DIR}* && \
+
+  ## [ux-diff] [cloud provider] UX differs across different Cloud providers
+  if [ "${SCENARIO_CLOUD_PROVIDER}" == "docker" ]; then
+
+    log "Environment is clean. Starting."
+
+  elif [ "${SCENARIO_CLOUD_PROVIDER}" == "aws" ]; then
+    ## [ ] [test] check if still works in AWS
+
+    ## [mark] this thing is ugly. Debug and check paths
+    log "Remove pre-existing configuration and stopping pre-existing Consul instances"
+    remote_exec consul-server-$i "sudo rm -rf ${CONSUL_CONFIG_DIR}* && \
                                   sudo mkdir -p ${CONSUL_CONFIG_DIR} && \
                                   sudo chown consul: ${CONSUL_CONFIG_DIR} && \
                                   sudo chmod g+w ${CONSUL_CONFIG_DIR} && \
@@ -53,12 +67,19 @@ for i in `seq 0 "$((SERVER_NUMBER-1))"`; do
                                   sudo chown consul: ${CONSUL_DATA_DIR} && \
                                   sudo chmod g+w ${CONSUL_DATA_DIR}"
   
-  _CONSUL_PID=`remote_exec consul-server-$i "pidof consul"`
-  if [ ! -z ${_CONSUL_PID} ]; then
-    remote_exec consul-server-$i "sudo kill -9 ${_CONSUL_PID}"
+    _CONSUL_PID=`remote_exec consul-server-$i "pidof consul"`
+    
+    if [ ! -z ${_CONSUL_PID} ]; then
+      remote_exec consul-server-$i "sudo kill -9 ${_CONSUL_PID}"
+    fi
+    
+  else 
+    log_err "Cloud provider $SCENARIO_CLOUD_PROVIDER is unsupported...exiting."
+    exit 245
   fi
-  
+
   log "Copying Configuration on consul-server-$i"
+  
   remote_copy consul-server-$i "${STEP_ASSETS}consul-server-$i/*" "${CONSUL_CONFIG_DIR}" 
 
 done
@@ -67,10 +88,15 @@ done
 ##########################################################
 header2 "Start Consul"
 
-## ~todo make all servers discoverable from bastion host
+## [x] todo make all servers discoverable from bastion host
 for i in `seq 0 "$((SERVER_NUMBER-1))"`; do
   log "Start Consul process on consul-server-$i"
   
+  ## [ux-diff] [cloud provider] UX differs across different Cloud providers
+  if [ "${SCENARIO_CLOUD_PROVIDER}" == "docker" ]; then
+    wait_for consul-server-$i
+  fi
+
   remote_exec consul-server-$i \
     "/usr/bin/consul agent \
     -log-file=/tmp/consul-server-$i.${DATACENTER}.${DOMAIN} \
@@ -78,7 +104,6 @@ for i in `seq 0 "$((SERVER_NUMBER-1))"`; do
 
   sleep 1
 done
-
 ##########################################################
 header2 "Configure ACL"
 
@@ -118,32 +143,3 @@ header2 "Configure servers token"
 
 ## [cmd] [script] generate_consul_server_tokens.sh
 execute_supporting_script "generate_consul_server_tokens.sh"
-
-##########################################################
-header2 "Change Server VMs DNS"
-
-## [feat] [flow] Change DNS for client
-## [ ] Check if it works on other Cloud providers
-## [ ] Check if it works as expected
-for i in `seq 0 "$((SERVER_NUMBER-1))"`; do
-  log "Change DNS configuration on consul-server-$i"
-  
-  _consul_resolv=$(cat << EOF
-
-domain ${CONSUL_DOMAIN}
-search ${CONSUL_DOMAIN}
-nameserver 127.0.0.1
-
-EOF
-)
-
-  # remote_exec consul-server-$i \
-  #   "echo -n \"${_consul_resolv}\n\" | cat - /etc/resolv.conf | sudo tee /etc/resolv.conf"
-  
-  remote_exec consul-server-$i \
-    "sudo iptables --table nat --append OUTPUT --destination localhost --protocol udp --match udp --dport 53 --jump REDIRECT --to-ports 8600 && \
-     sudo iptables --table nat --append OUTPUT --destination localhost --protocol tcp --match tcp --dport 53 --jump REDIRECT --to-ports 8600" 
-
-done
-
-
